refactor: 다중 필드 입력 공통 로직을 useMultiFieldInput으로 추출
